Somehow the player logout stuff got missed in a previous massive refactor. Oh well.. fixing it anyways!
