+++ src/main/java/org/spongepowered/common/event/tracking/phase/player/PlayerLogoutPhaseState.java
@@ -22,7 +22,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.spongepowered.common.event.tracking.phase;
+package org.spongepowered.common.event.tracking.phase.player;
 
 import org.spongepowered.api.Sponge;
 import org.spongepowered.api.entity.Entity;
@@ -31,65 +31,33 @@ import org.spongepowered.api.event.CauseStackManager;
 import org.spongepowered.common.entity.EntityUtil;
 import org.spongepowered.common.event.SpongeCommonEventFactory;
 import org.spongepowered.common.event.tracking.IPhaseState;
-import org.spongepowered.common.event.tracking.PhaseContext;
 import org.spongepowered.common.event.tracking.TrackingUtil;
 import org.spongepowered.common.event.tracking.context.GeneralizedContext;
+import org.spongepowered.common.event.tracking.phase.TrackingPhase;
+import org.spongepowered.common.event.tracking.phase.TrackingPhases;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
-public class PlayerPhase extends TrackingPhase {
+final class PlayerLogoutPhaseState implements IPhaseState<GeneralizedContext> {
 
-    public static final class State {
-
-        public static final IPhaseState<GeneralizedContext> PLAYER_LOGOUT = new PlayerPhaseState();
+    PlayerLogoutPhaseState() {
     }
 
-    static final class PlayerPhaseState implements IPhaseState<GeneralizedContext> {
-
-        PlayerPhaseState() {
-        }
-
-        @Override
-        public TrackingPhase getPhase() {
-            return TrackingPhases.PLAYER;
-        }
-
-        @Override
-        public GeneralizedContext createPhaseContext() {
-            return new GeneralizedContext(this);
-        }
-
-        @Override
-        public void unwind(GeneralizedContext phaseContext) {
-
-        }
-        private final String className = this.getClass().getSimpleName();
-
-        @Override
-        public String toString() {
-            return this.getPhase() + "{" + this.className + "}";
-        }
+    @Override
+    public TrackingPhase getPhase() {
+        return TrackingPhases.PLAYER;
     }
 
-    public static PlayerPhase getInstance() {
-        return Holder.INSTANCE;
+    @Override
+    public GeneralizedContext createPhaseContext() {
+        return new GeneralizedContext(this);
     }
 
-    private PlayerPhase() {
-    }
-
-    private static final class Holder {
-
-        static final PlayerPhase INSTANCE = new PlayerPhase();
-    }
-
-    @SuppressWarnings("unchecked")
-    public void unwind(IPhaseState<?> state, PhaseContext<?> phaseContext) {
-        // Since currently all we have is PLAYER_LOGOUT, don't care about
-        // states.
+    @Override
+    public void unwind(GeneralizedContext phaseContext) {
         final Player player = phaseContext.getSource(Player.class)
-                .orElseThrow(TrackingUtil.throwWithContext("Expected to be processing a player leaving, but we're not!", phaseContext));
+            .orElseThrow(TrackingUtil.throwWithContext("Expected to be processing a player leaving, but we're not!", phaseContext));
         try (CauseStackManager.StackFrame frame = Sponge.getCauseStackManager().pushCauseFrame()) {
             frame.pushCause(player);
             phaseContext.getCapturedItemsSupplier().acceptAndClearIfNotEmpty(items -> {
@@ -100,14 +68,19 @@ public class PlayerPhase extends TrackingPhase {
             });
             phaseContext.getCapturedItemStackSupplier().acceptAndClearIfNotEmpty(items -> {
                 final List<Entity> drops = items.stream()
-                        .map(drop -> drop.create(EntityUtil.getMinecraftWorld(player)))
-                        .map(EntityUtil::fromNative)
-                        .collect(Collectors.toList());
+                    .map(drop -> drop.create(EntityUtil.getMinecraftWorld(player)))
+                    .map(EntityUtil::fromNative)
+                    .collect(Collectors.toList());
                 SpongeCommonEventFactory.callDropItemCustom(drops, phaseContext);
             });
             phaseContext.getCapturedBlockSupplier()
-                    .acceptAndClearIfNotEmpty(blocks -> TrackingUtil.processBlockCaptures(blocks, state, phaseContext));
+                .acceptAndClearIfNotEmpty(blocks -> TrackingUtil.processBlockCaptures(blocks, this, phaseContext));
         }
     }
+    private final String className = this.getClass().getSimpleName();
 
+    @Override
+    public String toString() {
+        return this.getPhase() + "{" + this.className + "}";
+    }
 }
